Refactor read_config to happen on initialize, reactor specs
The specs now operate in the directory of the temp repositories.

Slight refactorings in the specs.

Add specs for the cli_options feature.

Change read_config to happen on the runner's initialize, so that certain
changes would be easier to test.

Change the way we get the repo_path (now pull from Rugged using the
current directory), so that it does not rely on the patch information.